Warning fixes

+++ examples/nodes/vsglights/vsglights.cpp
@@ -31,9 +31,9 @@ vsg::ref_ptr<vsg::Node> createTestScene(vsg::ref_ptr<vsg::Options> options)
 
     auto bounds = vsg::visit<vsg::ComputeBounds>(scene).bounds;
     double diameter = vsg::length(bounds.max - bounds.min);
-    geomInfo.position.set((bounds.min.x + bounds.max.x)*0.5, (bounds.min.y + bounds.max.y)*0.5, bounds.min.z);
-    geomInfo.dx.set(diameter, 0.0, 0.0);
-    geomInfo.dy.set(0.0, diameter, 0.0);
+    geomInfo.position.set(static_cast<float>((bounds.min.x + bounds.max.x)*0.5), static_cast<float>((bounds.min.y + bounds.max.y)*0.5), static_cast<float>(bounds.min.z));
+    geomInfo.dx.set(static_cast<float>(diameter), 0.0f, 0.0f);
+    geomInfo.dy.set(0.0f, static_cast<float>(diameter), 0.0f);
 
     scene->addChild(builder->createQuad(geomInfo, stateInfo));
 
@@ -127,8 +127,8 @@ int main(int argc, char** argv)
         {
             auto ambientLight = vsg::AmbientLight::create();
             ambientLight->name = "ambient";
-            ambientLight->color.set(1.0, 1.0, 1.0);
-            ambientLight->intensity = 0.01;
+            ambientLight->color.set(1.0f, 1.0f, 1.0f);
+            ambientLight->intensity = 0.01f;
             group->addChild(ambientLight);
         }
 
@@ -137,9 +137,9 @@ int main(int argc, char** argv)
         {
             auto directionalLight = vsg::DirectionalLight::create();
             directionalLight->name = "directional";
-            directionalLight->color.set(1.0, 1.0, 1.0);
-            directionalLight->intensity = 0.15;
-            directionalLight->direction.set(0.0, -1.0, -1.0);
+            directionalLight->color.set(1.0f, 1.0f, 1.0f);
+            directionalLight->intensity = 0.15f;
+            directionalLight->direction.set(0.0f, -1.0f, -1.0f);
             group->addChild(directionalLight);
         }
 
@@ -148,9 +148,9 @@ int main(int argc, char** argv)
         {
             auto pointLight = vsg::PointLight::create();
             pointLight->name = "point";
-            pointLight->color.set(1.0, 1.0, 0.0);
-            pointLight->intensity = span*0.5;
-            pointLight->position.set(bounds.min.x, bounds.min.y, bounds.max.z + span*0.3);
+            pointLight->color.set(1.0f, 1.0f, 0.0);
+            pointLight->intensity = static_cast<float>(span*0.5);
+            pointLight->position.set(static_cast<float>(bounds.min.x), static_cast<float>(bounds.min.y), static_cast<float>(bounds.max.z + span*0.3));
 
             // enable culling of the point light by decorating with a CullGroup
             auto cullGroup = vsg::CullGroup::create();
@@ -167,12 +167,12 @@ int main(int argc, char** argv)
         {
             auto spotLight = vsg::SpotLight::create();
             spotLight->name = "spot";
-            spotLight->color.set(0.0, 1.0, 1.0);
-            spotLight->intensity = span*0.5;
-            spotLight->position.set(bounds.max.x + span*0.1, bounds.min.y - span*0.1, bounds.max.z + span*0.3);
+            spotLight->color.set(0.0f, 1.0f, 1.0f);
+            spotLight->intensity = static_cast<float>(span*0.5);
+            spotLight->position.set(static_cast<float>(bounds.max.x + span*0.1), static_cast<float>(bounds.min.y - span*0.1), static_cast<float>(bounds.max.z + span*0.3));
             spotLight->direction = (bounds.min+bounds.max)*0.5 - spotLight->position;
-            spotLight->innerAngle = vsg::radians(8.0);
-            spotLight->outerAngle = vsg::radians(9.0);
+            spotLight->innerAngle = vsg::radians(8.0f);
+            spotLight->outerAngle = vsg::radians(9.0f);
 
             // enable culling of the spot light by decorating with a CullGroup
             auto cullGroup = vsg::CullGroup::create();
@@ -188,14 +188,14 @@ int main(int argc, char** argv)
         {
             auto ambientLight = vsg::AmbientLight::create();
             ambientLight->name = "ambient";
-            ambientLight->color.set(1.0, 1.0, 1.0);
-            ambientLight->intensity = 0.1;
+            ambientLight->color.set(1.0f, 1.0f, 1.0f);
+            ambientLight->intensity = 0.1f;
 
             auto directionalLight = vsg::DirectionalLight::create();
             directionalLight->name = "head light";
-            directionalLight->color.set(1.0, 1.0, 1.0);
-            directionalLight->intensity = 0.9;
-            directionalLight->direction.set(0.0, 0.0, -1.0);
+            directionalLight->color.set(1.0f, 1.0f, 1.0f);
+            directionalLight->intensity = 0.9f;
+            directionalLight->direction.set(0.0f, 0.0f, -1.0f);
 
             auto absoluteTransform = vsg::AbsoluteTransform::create();
             absoluteTransform->addChild(ambientLight);
